Separated Placement and Inventory
