+++ src/RealmClient/Assets/_Scripts/DatabaseController.cs
@@ -1,6 +1,8 @@
 using UnityEngine;
 using PocketBaseSdk;
 using UnityEngine.SceneManagement;
+using System.Threading.Tasks;
+using System.Collections.Generic;
 public class DatabaseController : MonoBehaviour
 {
     // Start is called once before the first execution of Update after the MonoBehaviour is created
@@ -15,9 +17,31 @@ public class DatabaseController : MonoBehaviour
         }
     }
 
-    public void logout()
+    public void Logout()
     {
         pb.AuthStore.Clear();
         SceneManager.LoadScene(0);
+    }
+
+    public async Task<List<AnchorDTO>> GetAllARObjects()
+    {
+        var rawAnchors = await pb.Collection("ar_objects").GetFullList();
+        List<AnchorDTO> anchors = new List<AnchorDTO>();
+        foreach (RecordModel a in rawAnchors)
+        {
+            anchors.Add(AnchorDTO.FromRecord(a));
+        }
+        return anchors;
+    }
+
+    public async Task<AnchorDTO> GetARObjectFromAnchorId(string anchorId)
+    {
+        var anchors = await GetAllARObjects();
+        foreach (AnchorDTO anchorDTO in anchors)
+        {
+            if (anchorDTO.AnchorId == anchorId)
+                return anchorDTO;
+        }
+        return new AnchorDTO();
     }
 }
